Move tests to correct folder
